Some detail page code cleanup/consolidation - new HealthStatusPopover (consolidate renderHealthStatus instances) - consolidate repeated styles - consolidate repeated label click-filtering logic
