+++ frontend/src/pages/AppDetails/AppDetailsPage.tsx
@@ -29,26 +29,7 @@ import { isGVKSupported } from '../../utils/IstioConfigUtils';
 import { getAppLabelName } from 'config/ServerConfig';
 import { setAIContext } from 'helpers/ChatAI';
 import { PFBadge, PFBadges } from '../../components/Pf/PfBadges';
-import { kialiStyle } from 'styles/StyleUtils';
-
-const titleRowStyle = kialiStyle({
-  alignItems: 'center',
-  display: 'flex',
-  flexWrap: 'nowrap',
-  gap: 'var(--pf-t--global--spacer--md)',
-  marginTop: '0.5rem',
-  minWidth: 0,
-  width: '100%'
-});
-
-const titleMainStyle = kialiStyle({
-  alignItems: 'center',
-  display: 'flex',
-  flex: '1 1 auto',
-  flexWrap: 'nowrap',
-  gap: 'var(--pf-t--global--spacer--sm)',
-  minWidth: 0
-});
+import { detailTitleRowStyle, detailTitleMainStyle } from 'styles/FlexStyles';
 
 type AppDetailsState = {
   app?: App;
@@ -320,8 +301,8 @@ class AppDetails extends React.Component<AppDetailsProps, AppDetailsState> {
       <>
         <RenderHeader rightToolbar={<TimeControl customDuration={useCustomTime} />}>
           {this.state.app && (
-            <div className={titleRowStyle}>
-              <div className={titleMainStyle}>
+            <div className={detailTitleRowStyle}>
+              <div className={detailTitleMainStyle}>
                 <PFBadge badge={PFBadges.App} position={TooltipPosition.top} />
                 <Title headingLevel="h1" size={TitleSizes.xl} style={{ margin: 0, flexShrink: 0 }}>
                   {this.props.appId.app}
