+++ frontend/src/pages/NamespaceDetails/NamespaceDetailsPage.tsx
@@ -8,6 +8,7 @@ import { NamespaceInfo } from 'types/NamespaceInfo';
 import { PromisesRegistry } from 'utils/CancelablePromises';
 import * as API from 'services/Api';
 import { addError, addSuccess } from 'utils/AlertUtils';
+import { buildMetadataPatch } from '../PageUtils';
 import { ParameterizedTabs, activeTab } from 'components/Tab/Tabs';
 import { RenderHeader } from 'components/Nav/Page/RenderHeader';
 import { ErrorSection } from 'components/ErrorSection/ErrorSection';
@@ -34,29 +35,10 @@ import { MessageType } from 'types/NotificationCenter';
 import { setControlPlaneRevisions } from 'pages/Namespaces/NamespaceRevisionUtils';
 import { PFBadge, PFBadges } from 'components/Pf/PfBadges';
 import { TimeControl } from 'components/Time/TimeControl';
-import { kialiStyle } from 'styles/StyleUtils';
+import { detailTitleRowStyle, detailTitleMainStyle } from 'styles/FlexStyles';
 import { RefreshIntervalManual } from 'config/Config';
 import { serverConfig } from 'config/ServerConfig';
 import { t } from 'utils/I18nUtils';
-
-const titleRowStyle = kialiStyle({
-  alignItems: 'center',
-  display: 'flex',
-  flexWrap: 'nowrap',
-  gap: 'var(--pf-t--global--spacer--md)',
-  marginTop: '0.5rem',
-  minWidth: 0,
-  width: '100%'
-});
-
-const titleMainStyle = kialiStyle({
-  alignItems: 'center',
-  display: 'flex',
-  flex: '1 1 auto',
-  flexWrap: 'nowrap',
-  gap: 'var(--pf-t--global--spacer--sm)',
-  minWidth: 0
-});
 
 type ReduxProps = {
   duration: DurationInSeconds;
@@ -404,13 +386,7 @@ export class NamespaceDetailsPageComponent extends React.Component<NamespaceDeta
 
   private handleSaveMetadata = (field: 'labels' | 'annotations', updated: Record<string, string>): void => {
     const original = (field === 'labels' ? this.state.nsInfo?.labels : this.state.nsInfo?.annotations) ?? {};
-    const patch: Record<string, string | null> = { ...updated };
-    Object.keys(original).forEach(key => {
-      if (!(key in updated)) {
-        patch[key] = null;
-      }
-    });
-    const jsonPatch = JSON.stringify({ metadata: { [field]: patch } });
+    const jsonPatch = buildMetadataPatch(field, original, updated);
 
     API.updateNamespace(this.props.namespace, jsonPatch, this.state.cluster)
       .then(() => {
@@ -452,8 +428,8 @@ export class NamespaceDetailsPageComponent extends React.Component<NamespaceDeta
           rightToolbar={<TimeControl customDuration={false} />}
         >
           {!this.state.error && ns && (
-            <div className={titleRowStyle} data-test="namespace-detail-title-row">
-              <div className={titleMainStyle}>
+            <div className={detailTitleRowStyle} data-test="namespace-detail-title-row">
+              <div className={detailTitleMainStyle}>
                 <PFBadge badge={PFBadges.Namespace} position={TooltipPosition.top} />
                 <Title headingLevel="h1" size={TitleSizes.xl} style={{ margin: 0, flexShrink: 0 }}>
                   {this.props.namespace}
