+++ frontend/src/pages/WorkloadDetails/WorkloadDetailsPage.tsx
@@ -40,26 +40,7 @@ import { WaypointConfig } from '../../components/Ambient/WaypointConfig';
 import { isGVKSupported } from '../../utils/IstioConfigUtils';
 import { setAIContext } from 'helpers/ChatAI';
 import { PFBadge, PFBadges } from '../../components/Pf/PfBadges';
-import { kialiStyle } from 'styles/StyleUtils';
-
-const titleRowStyle = kialiStyle({
-  alignItems: 'center',
-  display: 'flex',
-  flexWrap: 'nowrap',
-  gap: 'var(--pf-t--global--spacer--md)',
-  marginTop: '0.5rem',
-  minWidth: 0,
-  width: '100%'
-});
-
-const titleMainStyle = kialiStyle({
-  alignItems: 'center',
-  display: 'flex',
-  flex: '1 1 auto',
-  flexWrap: 'nowrap',
-  gap: 'var(--pf-t--global--spacer--sm)',
-  minWidth: 0
-});
+import { detailTitleRowStyle, detailTitleMainStyle } from 'styles/FlexStyles';
 
 type WorkloadDetailsState = {
   cluster?: string;
@@ -449,8 +430,8 @@ class WorkloadDetailsPageComponent extends React.Component<WorkloadDetailsPagePr
           actionsToolbarTop="11.1rem"
         >
           {this.state.workload && (
-            <div className={titleRowStyle}>
-              <div className={titleMainStyle}>
+            <div className={detailTitleRowStyle}>
+              <div className={detailTitleMainStyle}>
                 <PFBadge badge={PFBadges.Workload} position={TooltipPosition.top} />
                 <Title headingLevel="h1" size={TitleSizes.xl} style={{ margin: 0, flexShrink: 0 }}>
                   {this.state.workload.name}
